streamlit/requirements.txt
Pillow==9.4.0
Pillow==10.3.0
python_docx==1.1.0
Requests==2.31.0
streamlit==1.33.0
